fix(firestore): update sample getting_started_sessions_setup to make a direct implementation of fs client
